Add `validate_args=True` to distributions in tests.
Fix a handful of bugs this exposed:
- Consistent dtype in assertions in `Beta` and `Categorical`
- Use `tensorshape_util` to check unknown batch shapes in `HiddenMarkovModel`
- `scale` must be positive definite in `StudentTProcess`
- `scale` is nonsingular in `VariationalGaussianProcess`
- Fixed assertions in `Triangular`

PiperOrigin-RevId: 280045729

+++ tensorflow_probability/python/distributions/categorical.py
@@ -276,7 +276,7 @@ class Categorical(distribution.Distribution):
     logits = self.logits_parameter()
     if self.validate_args:
       k = distribution_util.embed_check_integer_casting_closed(
-          k, target_dtype=tf.int32)
+          k, target_dtype=self.dtype)
     k, logits = _broadcast_cat_event_and_params(
         k, logits, base_dtype=dtype_util.base_dtype(self.dtype))
     return -tf.nn.sparse_softmax_cross_entropy_with_logits(
